refactor(predict): organize event card module
