+++ app/components/UI/PredictNext/AGENTS.md
@@ -12,7 +12,8 @@ Read, in order:
 4. `docs/architecture.md` — stable boundaries and invariants.
 5. `docs/canonical-read-model-and-api.md` — agreed working direction for public read types and routes.
 6. `docs/ui-components.md` — composition rules for PredictNext core UI work.
-7. The Jira issue being implemented.
+7. `docs/module-structure.md` — feature ownership and public/internal module boundaries.
+8. The Jira issue being implemented.
 
 ## Scope
 
@@ -41,6 +42,7 @@ Proposed ADRs and examples in these docs are working direction, not accepted con
 - Never put credentials, bearer tokens, OTPs, PII/KYC values, or transfer-authorization material in Redux, persisted mobile storage, logs, analytics, traces, or fixtures.
 - Follow repository TypeScript, testing, design-system, and security guidance.
 - For core UI, follow `docs/ui-components.md`: prefer explicit variants and children-based composition, avoid boolean mode props, and use compound components only where independently useful parts or shared state justify them.
+- For placement and imports, follow `docs/module-structure.md`: organize product UI by owning feature/domain, import across modules through deliberate `index.ts` APIs, and keep private implementation under the owner's `internal/` directory.
 - Work one behavior at a time: red → green → refactor.
 
 ## Agent readiness
