+++ app/components/UI/PredictNext/events/cards/index.ts
@@ -1,0 +1,4 @@
+export { EventCardGame } from './EventCardGame';
+export type { EventCardGameProps, EventCardGameVariant } from './EventCardGame';
+export { EventCardStandard } from './EventCardStandard';
+export type { EventCardStandardProps } from './EventCardStandard';
