+++ app/components/UI/PredictNext/events/cards/internal/formatAskPrice.test.ts
@@ -1,4 +1,4 @@
-import type { PredictDecimal } from '../../types';
+import type { PredictDecimal } from '../../../types';
 import { formatAskPrice } from './formatAskPrice';
 
 describe('formatAskPrice', () => {
