+++ app/components/UI/PredictNext/events/cards/internal/formatMultiplier.test.ts
@@ -1,4 +1,4 @@
-import type { PredictDecimal } from '../../types';
+import type { PredictDecimal } from '../../../types';
 import { formatMultiplier } from './formatMultiplier';
 
 describe('formatMultiplier', () => {
